alerts nos gerentes

+++ lib/views/managers_view.dart
@@ -146,7 +146,68 @@ class ManagersView extends StatelessWidget {
                                 icon: const Icon(Icons.delete),
                                 color: const Color.fromARGB(255, 255, 0, 0),
                                 onPressed: () {
-                                  state.deleteManagers(manager);
+                                  showDialog(
+                                    context: context,
+                                    builder: (context) {
+                                      return AlertDialog(
+                                        title: const Padding(
+                                          padding: EdgeInsets.symmetric(
+                                            horizontal: 5,
+                                            vertical: 20,
+                                          ),
+                                          child: Text(
+                                            'Excluir Gerente',
+                                            style: TextStyle(
+                                              color: Colors.black,
+                                              fontWeight: FontWeight.bold,
+                                              fontSize: 25.0,
+                                            ),
+                                          ),
+                                        ),
+                                        content: const Text(
+                                          'Deseja excluir este gerente?',
+                                          style: TextStyle(
+                                            color: Colors.black,
+                                            fontSize: 20.0,
+                                          ),
+                                        ),
+                                        actions: [
+                                          TextButton(
+                                            onPressed: () {
+                                              Navigator.pop(context);
+                                            },
+                                            child: const Text(
+                                              'Cancelar',
+                                              style: TextStyle(
+                                                color: Colors.red,
+                                                fontWeight: FontWeight.bold,
+                                                fontSize: 20.0,
+                                              ),
+                                            ),
+                                          ),
+                                          TextButton(
+                                            onPressed: () {
+                                              state.deleteManagers(manager);
+                                              Navigator.of(context).pop();
+                                            },
+                                            child: const Text(
+                                              'Confirmar',
+                                              style: TextStyle(
+                                                color: Colors.black,
+                                                fontSize: 20.0,
+                                              ),
+                                            ),
+                                          ),
+                                        ],
+                                        elevation: 25,
+                                        backgroundColor: Colors.white,
+                                        shape: RoundedRectangleBorder(
+                                          borderRadius:
+                                              BorderRadius.circular(50.0),
+                                        ),
+                                      );
+                                    },
+                                  );
                                 },
                               ),
                               IconButton(
